fix: 필요없는 코드 삭제 및 파이프라인 정리

+++ VPGraphics/ModelLoader.cpp
@@ -202,7 +202,7 @@ void ModelLoader::ProcessMesh(std::shared_ptr<ModelData> Model, aiMesh* mesh, un
 
 				DirectX::XMFLOAT4X4 temp;
 
-				for (int j = 0; j < curAiBone->mNumWeights; j++)
+				for (int j = 0; j < static_cast<int>(curAiBone->mNumWeights); j++)
 				{
 					curBone->weights.push_back(curAiBone->mWeights[j].mWeight);
 					curBone->vertexids.push_back(curAiBone->mWeights[j].mVertexId);
@@ -624,7 +624,7 @@ void ModelLoader::ProcessBoneMapping(std::vector<SkinningVertex>& buffer, aiMesh
 				{
 					if (buffer[curVertexId].BoneIndices[j] == 0.0f)
 					{
-						buffer[curVertexId].BoneIndices[j] = i;
+						buffer[curVertexId].BoneIndices[j] = static_cast<float>(i);
 						break;
 					}
 				}
